super-agent-langgraph/requirements.txt
langgraph>=0.4.7
langchain>=0.3.25
langchain-openai>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.1
aiohttp>=3.12.6
python-dotenv>=1.1.0
pydantic>=2.11.5
pydantic-settings>=2.9.1
